Almost done

+++ lms-ui/src/components/content/Navbar.js
@@ -3,8 +3,9 @@ import user from "./Images/user.png";
 import "../styles/HomePage.css";
 import { Link, useNavigate } from "react-router-dom";
 import Avatar from "react-avatar";
+import Swal from "sweetalert2";
 
-export const Navbar = () => {
+export const Navbar = ({ page }) => {
   const navigate = useNavigate();
   const name = `${sessionStorage.getItem(
     "first_name"
@@ -12,8 +13,20 @@ export const Navbar = () => {
   const role = sessionStorage.getItem("role");
 
   const handleLogout = () => {
-    sessionStorage.clear();
-    navigate("/auth/login");
+    Swal.fire({
+      title: "Attention!",
+      text: `Are You Sure?`,
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: `SignOut`,
+    }).then((result) => {
+      if (result.isConfirmed) {
+        sessionStorage.clear();
+        navigate("/auth/login");
+      }
+    });
   };
 
   function randomColor() {
@@ -26,11 +39,11 @@ export const Navbar = () => {
   return (
     <nav
       className="navbar navbar-expand-md mb-0 nav-bar p-0"
-      style={{ height: "8%", boxShadow: "4px 3px 21px -10px gray" }}
+      style={{ boxShadow: "4px 3px 21px -10px gray" }}
     >
       <div className="container-fluid">
-        <Link to="/" className="navbar-brand" href="#">
-          Dashboard
+        <Link to="" className="navbar-brand">
+          { page && page.charAt(0).toUpperCase() + page.slice(1)}
         </Link>
         <button
           className="navbar-toggler"
@@ -55,17 +68,17 @@ export const Navbar = () => {
             <input type="search" placeholder="Search" />
           </div> */}
           <ul className="navbar-nav ml-auto navbar-right ">
-            <li className="nav-item">
+            {/* <li className="nav-item">
               <a
                 className="nav-link notification-icon"
-                href="/"
+                // href="/"
                 data-target="#myModal"
                 data-toggle="modal"
               >
                 <i className="fas fa-sharp fa-solid fa-message-sms"></i>{" "}
               </a>
-            </li>
-            <li className="nav-item">
+            </li> */}
+            {/* <li className="nav-item">
               <a
                 className="nav-link notification-icon"
                 href="/"
@@ -74,11 +87,11 @@ export const Navbar = () => {
               >
                 <i className="fas fa-solid fa-bell"></i>
               </a>
-            </li>
+            </li> */}
             <li className="nav-item">
               <a
                 className="nav-link"
-                href="/"
+                href="#"
                 data-target="#myModal"
                 data-toggle="modal"
               >
@@ -96,7 +109,7 @@ export const Navbar = () => {
             <li className="nav-item">
               <a
                 className="nav-link"
-                href="/"
+                href="#"
                 data-target="#myModal"
                 data-toggle="modal"
               >
@@ -108,13 +121,17 @@ export const Navbar = () => {
             </li>
             <li>
               <div className="logout-container">
-                <button
+                {/* <button
                   className="log-out-btn"
                   type="button"
                   onClick={() => handleLogout()}
-                >
-                  LogOut
-                </button>
+                > */}
+                <i
+                  class="bi bi-box-arrow-right log-out-btn"
+                  type="button"
+                  onClick={() => handleLogout()}
+                ></i>
+                {/* </button> */}
               </div>
             </li>
           </ul>
